requirements.txt
flask
flask-cors
python-dotenv
pandas
sentence-transformers
torch
transformers
scikit-learn
google-generativeai
gunicorn
Flask==3.0.3
Flask-Cors==5.0.0
pandas==2.2.2
sentence-transformers==3.0.1
torch==2.4.1
google-generativeai==0.7.2
python-dotenv==1.0.1
gunicorn==23.0.0
